task_1/requirements-dev.txt
-r requirements.txt
pytest
django-stubs
djangorestframework-stubs
